:recycle: 채팅방 컨트롤러 코드 정리
