+++ src/main/kotlin/com/stark/shoot/adapter/in/web/chatroom/ChatRoomController.kt
@@ -2,6 +2,7 @@ package com.stark.shoot.adapter.`in`.web.chatroom
 
 import com.stark.shoot.adapter.`in`.web.dto.ResponseDto
 import com.stark.shoot.adapter.`in`.web.dto.chatroom.ChatRoomResponse
+import com.stark.shoot.adapter.`in`.web.dto.chatroom.TitleRequest
 import com.stark.shoot.application.port.`in`.chatroom.CreateChatRoomUseCase
 import com.stark.shoot.application.port.`in`.chatroom.FindChatRoomUseCase
 import com.stark.shoot.application.port.`in`.chatroom.ManageChatRoomUseCase
@@ -13,9 +14,9 @@ import org.springframework.web.bind.annotation.*
 @RestController
 @RequestMapping("/api/v1/chatrooms")
 class ChatRoomController(
+    private val findChatRoomUseCase: FindChatRoomUseCase,
     private val createChatRoomUseCase: CreateChatRoomUseCase,
-    private val manageChatRoomUseCase: ManageChatRoomUseCase,
-    private val findChatRoomUseCase: FindChatRoomUseCase
+    private val manageChatRoomUseCase: ManageChatRoomUseCase
 ) {
 
     @Operation(summary = "1:1 채팅방 생성", description = "특정 사용자와 친구의 1:1 채팅방을 생성합니다.")
@@ -28,6 +29,15 @@ class ChatRoomController(
         return ResponseDto.success(ChatRoomResponse.from(room, userId), "채팅방이 생성되었습니다.")
     }
 
+    @Operation(summary = "사용자의 채팅방 목록 조회", description = "특정 사용자의 채팅방 전체 목록을 조회합니다.")
+    @GetMapping
+    fun getChatRooms(
+        @RequestParam userId: Long
+    ): ResponseDto<List<ChatRoomResponse>> {
+        val chatRooms = findChatRoomUseCase.getChatRoomsForUser(userId)
+        return ResponseDto.success(chatRooms)
+    }
+
     @Operation(summary = "채팅방 퇴장", description = "현재 사용자가 채팅방에서 퇴장합니다.")
     @DeleteMapping("/{roomId}/exit")
     fun exitChatRoom(
@@ -38,21 +48,14 @@ class ChatRoomController(
         return ResponseDto.success(result, "채팅방에서 퇴장했습니다.")
     }
 
-    @Operation(
-        summary = "두 사용자 간의 1:1 채팅방 찾기",
-        description = "두 사용자 ID를 받아 해당 사용자들 간의 1:1 채팅방을 찾습니다."
-    )
-    @GetMapping("/direct")
-    fun findDirectChatRoom(
-        @RequestParam myId: Long,
-        @RequestParam otherUserId: Long
-    ): ResponseDto<ChatRoomResponse> {
-        val chatRoom = findChatRoomUseCase.findDirectChatBetweenUsers(myId, otherUserId)
-        return if (chatRoom != null) {
-            ResponseDto.success(chatRoom, "채팅방을 찾았습니다.")
-        } else {
-            ResponseDto.fail("채팅방을 찾을 수 없습니다.", 404)
-        }
+    @Operation(summary = "채팅방 제목 변경", description = "채팅방의 제목을 변경합니다.")
+    @PutMapping("/{roomId}/title")
+    fun updateTitle(
+        @PathVariable roomId: Long,
+        @RequestBody request: TitleRequest
+    ): ResponseDto<Boolean> {
+        val result = manageChatRoomUseCase.updateTitle(roomId, request.title)
+        return ResponseDto.success(result, "채팅방 제목이 변경되었습니다.")
     }
 
 }
