+++ src/main/kotlin/com/stark/shoot/adapter/in/web/chatroom/ChatRoomListController.kt
@@ -17,16 +17,6 @@ class ChatRoomListController(
     private val findChatRoomUseCase: FindChatRoomUseCase
 ) {
 
-    @Operation(
-        summary = "사용자의 채팅방 목록 조회",
-        description = "특정 사용자의 채팅방 전체 목록을 조회합니다."
-    )
-    @GetMapping
-    fun getChatRooms(
-        @RequestParam userId: Long
-    ): ResponseDto<List<ChatRoomResponse>> {
-        val chatRooms = findChatRoomUseCase.getChatRoomsForUser(userId)
-        return ResponseDto.success(chatRooms)
-    }
+
 
 }
