+++ src/main/kotlin/com/stark/shoot/adapter/in/web/chatroom/ChatRoomSearchController.kt
@@ -3,7 +3,7 @@ package com.stark.shoot.adapter.`in`.web.chatroom
 import com.stark.shoot.adapter.`in`.web.dto.ResponseDto
 import com.stark.shoot.adapter.`in`.web.dto.chatroom.ChatRoomResponse
 import com.stark.shoot.application.port.`in`.chatroom.ChatRoomSearchUseCase
-import com.stark.shoot.infrastructure.util.toObjectId
+import com.stark.shoot.application.port.`in`.chatroom.FindChatRoomUseCase
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.web.bind.annotation.GetMapping
@@ -15,7 +15,8 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 @RequestMapping("/api/v1/chatrooms")
 class ChatRoomSearchController(
-    private val chatRoomSearchUseCase: ChatRoomSearchUseCase
+    private val chatRoomSearchUseCase: ChatRoomSearchUseCase,
+    private val findChatRoomUseCase: FindChatRoomUseCase
 ) {
 
     @Operation(summary = "채팅방 검색", description = "채팅방을 검색합니다.")
@@ -28,6 +29,23 @@ class ChatRoomSearchController(
     ): ResponseDto<List<ChatRoomResponse>> {
         val results = chatRoomSearchUseCase.searchChatRooms(userId, query, type, unreadOnly)
         return ResponseDto.success(results)
+    }
+
+    @Operation(
+        summary = "두 사용자 간의 1:1 채팅방 찾기",
+        description = "두 사용자 ID를 받아 해당 사용자들 간의 1:1 채팅방을 찾습니다."
+    )
+    @GetMapping("/direct")
+    fun findDirectChatRoom(
+        @RequestParam myId: Long,
+        @RequestParam otherUserId: Long
+    ): ResponseDto<ChatRoomResponse> {
+        val chatRoom = findChatRoomUseCase.findDirectChatBetweenUsers(myId, otherUserId)
+        return if (chatRoom != null) {
+            ResponseDto.success(chatRoom, "채팅방을 찾았습니다.")
+        } else {
+            ResponseDto.fail("채팅방을 찾을 수 없습니다.", 404)
+        }
     }
 
 }
